web/requirements.txt
flask_login
flask_bcrypt
pyodbc
pandas
flask_wtf
scikit-learn
email-validator
flask>=2.2.2
flask-sqlalchemy>=3.0.0
psycopg2>=2.9.4
gunicorn>=20.1.0
python-dotenv
watchdog